rust: Split interface generation to a separate file
